JCLOUDS-717: Join Enterprise and OpenSource Chef
