+++ apis/chef/src/main/java/org/jclouds/chef/filters/SignedHeaderAuth.java
@@ -63,7 +63,7 @@ import com.google.common.io.ByteSource;
 /**
  * Ported from mixlib-authentication in order to sign Chef requests.
  * 
- * @see <a href= "http://github.com/opscode/mixlib-authentication" />
+ * @see <a href= "https://github.com/chef/mixlib-authentication" />
  */
 @Singleton
 public class SignedHeaderAuth implements HttpRequestFilter {
