+++ apis/chef/src/test/java/org/jclouds/chef/binders/BindGroupToUpdateRequestJsonPayloadTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jclouds.enterprisechef.binders;
+package org.jclouds.chef.binders;
 
 import static org.testng.Assert.assertEquals;
 
@@ -23,7 +23,7 @@ import java.net.URI;
 
 import org.jclouds.chef.ChefApiMetadata;
 import org.jclouds.chef.config.ChefParserModule;
-import org.jclouds.enterprisechef.domain.Group;
+import org.jclouds.chef.domain.Group;
 import org.jclouds.http.HttpRequest;
 import org.jclouds.json.config.GsonModule;
 import org.jclouds.rest.annotations.ApiVersion;
