+++ apis/chef/src/test/java/org/jclouds/chef/features/OrganizationApiLiveTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jclouds.enterprisechef;
+package org.jclouds.chef.features;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
@@ -25,98 +25,96 @@ import static org.testng.Assert.assertTrue;
 import java.util.Set;
 import java.util.UUID;
 
-import org.jclouds.chef.internal.BaseChefApiLiveTest;
-import org.jclouds.enterprisechef.domain.Group;
-import org.jclouds.enterprisechef.domain.User;
+import org.jclouds.chef.domain.Group;
+import org.jclouds.chef.domain.User;
+import org.jclouds.chef.internal.BaseChefLiveTest;
 import org.jclouds.rest.ResourceNotFoundException;
+import org.testng.SkipException;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 /**
- * Tests behavior of the EnterpriseChefApi.
+ * Tests behavior of the OrganizationApi.
  */
-@Test(groups = "live", singleThreaded = true, testName = "EnterpriseChefApiLiveTest")
-public class EnterpriseChefApiLiveTest extends BaseChefApiLiveTest<EnterpriseChefApi> {
-
+@Test(groups = "live", singleThreaded = true, testName = "OrganizationApiLiveTest")
+public class OrganizationApiLiveTest extends BaseChefLiveTest {
    private static final String GROUP_NAME = System.getProperty("user.name") + "-jcloudstest";
-   private static final String ORG_NAME = System.getProperty("test.enterprisechef.org");
 
-   public EnterpriseChefApiLiveTest() {
-      provider = "enterprisechef";
-   }
+   private OrganizationApi orgApi;
 
-   @Override
-   @Test
-   public void testSearchClientsWithOptions() throws Exception {
-      // This test will fail because Enterprise Chef does not index client name.
-      // Once it is fixes, the test should succeed.
-      // See: http://tickets.opscode.com/browse/CHEF-2477
-      super.testSearchClientsWithOptions();
+   @BeforeMethod
+   public void skipIfApiNotAvailable() {
+      // Throwing SkipExceptions only works with @Test and @BeforeMethod methods
+      if (!api.organizationApi().isPresent()) {
+         throw new SkipException("Organization api not available in this Chef version");
+      }
+      orgApi = api.organizationApi().get();
    }
 
    public void testGetUser() {
-      User user = api.getUser(identity);
+      User user = orgApi.getUser(identity);
       assertEquals(user.getUsername(), identity);
       assertNotNull(user.getPublicKey());
    }
 
    public void testGetUnexistingUser() {
-      User user = api.getUser(UUID.randomUUID().toString());
+      User user = orgApi.getUser(UUID.randomUUID().toString());
       assertNull(user);
    }
 
    public void testListGroups() {
-      Set<String> groups = api.listGroups();
+      Set<String> groups = orgApi.listGroups();
       assertNotNull(groups);
       assertFalse(groups.isEmpty());
    }
 
    public void testGetUnexistingGroup() {
-      Group group = api.getGroup(UUID.randomUUID().toString());
+      Group group = orgApi.getGroup(UUID.randomUUID().toString());
       assertNull(group);
    }
 
    public void testCreateGroup() {
-      api.createGroup(GROUP_NAME);
-      Group group = api.getGroup(GROUP_NAME);
+      orgApi.createGroup(GROUP_NAME);
+      Group group = orgApi.getGroup(GROUP_NAME);
       assertNotNull(group);
       assertEquals(group.getGroupname(), GROUP_NAME);
    }
 
    @Test(dependsOnMethods = "testCreateGroup")
    public void testUpdateGroup() {
-      Group group = api.getGroup(GROUP_NAME);
+      Group group = orgApi.getGroup(GROUP_NAME);
       Group updated = Group.builder(group.getGroupname()) //
             .actors(group.getActors()) //
             .orgname(group.getOrgname()) //
             .name(group.getName()) //
             .groups(group.getGroups()) //
-            .client(ORG_NAME + "-validator") //
+            .client(group.getOrgname() + "-validator") //
             .user(identity) //
             .build();
 
-      api.updateGroup(updated);
-      group = api.getGroup(GROUP_NAME);
+      orgApi.updateGroup(updated);
+      group = orgApi.getGroup(GROUP_NAME);
 
       assertNotNull(group);
       assertTrue(group.getUsers().contains(identity));
-      assertTrue(group.getClients().contains(ORG_NAME + "-validator"));
+      assertTrue(group.getClients().contains(group.getOrgname() + "-validator"));
    }
 
    @Test(expectedExceptions = ResourceNotFoundException.class)
    public void testUpdateUnexistingGroup() {
-      api.updateGroup(Group.builder(UUID.randomUUID().toString()).build());
+      orgApi.updateGroup(Group.builder(UUID.randomUUID().toString()).build());
    }
 
-   @Test(dependsOnMethods = "testUpdateGroup")
+   @Test(dependsOnMethods = "testUpdateGroup", alwaysRun = true)
    public void testDeleteGroup() {
-      api.deleteGroup(GROUP_NAME);
-      Group group = api.getGroup(GROUP_NAME);
+      orgApi.deleteGroup(GROUP_NAME);
+      Group group = orgApi.getGroup(GROUP_NAME);
       assertNull(group);
    }
 
    @Test(expectedExceptions = ResourceNotFoundException.class)
    public void testDeleteUnexistingGroup() {
-      api.deleteGroup(UUID.randomUUID().toString());
+      orgApi.deleteGroup(UUID.randomUUID().toString());
    }
 
 }
