+++ apis/chef/src/test/java/org/jclouds/chef/strategy/internal/CleanupStaleNodesAndClientsImplLiveTest.java
@@ -19,7 +19,6 @@ package org.jclouds.chef.strategy.internal;
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertNull;
 
-import org.jclouds.chef.ChefApi;
 import org.jclouds.chef.internal.BaseChefLiveTest;
 import org.testng.annotations.Test;
 
@@ -29,7 +28,7 @@ import com.google.common.collect.ImmutableSet;
  * Tests behavior of {@code CleanupStaleNodesAndClientsImpl} strategies
  */
 @Test(groups = "live", testName = "CleanupStaleNodesAndClientsImplLiveTest")
-public class CleanupStaleNodesAndClientsImplLiveTest extends BaseChefLiveTest<ChefApi> {
+public class CleanupStaleNodesAndClientsImplLiveTest extends BaseChefLiveTest {
 
    private CreateNodeAndPopulateAutomaticAttributesImpl creator;
    private CleanupStaleNodesAndClientsImpl strategy;
