+++ apis/chef/src/test/resources/group-update.json
