+++ apis/chef/src/test/resources/group.json
