+++ apis/chef/src/test/resources/groups.json
