+++ apis/chef/src/test/resources/user.json
@@ -1,6 +1,6 @@
 {
   "display_name" : "Ignasi Barrera",
-  "email" : "myemail@enterprisechef.org",
+  "email" : "myemail@chef.org",
   "first_name" : "Ignasi",
   "last_name" : "Barrera",
   "middle_name" : "",
