+++ providers/pom.xml
@@ -72,6 +72,5 @@
     <module>aws-route53</module>
     <module>ultradns-ws</module>
     <module>dynect</module>
-    <module>enterprisechef</module>
   </modules>
 </project>
